+++ CHANGELOG.md
@@ -9,8 +9,10 @@ The format is based on the [KeepAChangeLog] project.
 
 ### Removed
 - [#671] Removed deprecated request/response_cls kwargs from Provider/Client methods
+- [#677] Removed more deprecated code
 
 [#671]: https://github.com/OpenIDC/pyoidc/pull/671
+[#677]: https://github.com/OpenIDC/pyoidc/pull/677
 
 ## 1.0.1 [2019-06-30]